Remove cyclic dependency between fetchApi.ts and src/index.ts (#84)
* move EdgeRequest and EdgeResponse to fetchApi.ts to avoid co-dependency between index.ts and fetchApi.ts

* move EdgeRequest and EdgeResponse to its new file
